requirements.txt
# === Agent + LLM Tools ===

aisuite==0.1.11

anthropic

docstring-parser

markdown

mistralai

openai

qrcode

tavily-python>=0.7.12

textstat

vertexai



# === Web Framework + API ===

fastapi

pydantic

pydantic[email]

python-dotenv

python-multipart

requests

sqlalchemy

uvicorn



# === Notebook Experience ===

ipywidgets

jupyter_server

nbclassic

notebook



# === Data Analysis / Display (Optional Enhancements) ===

duckdb

matplotlib

pandas

seaborn

tabulate

tinydb




# === Machine Learning / NLP (Optional Enhancements) ===

jinja2

psycopg2-binary

scikit-learn

Wikipedia

# === AI LLM
langgraph

langchain

langchain-openai